requirements.txt
numpy==1.20.1
pandas==1.2.3
python-dateutil==2.8.1
pytz==2021.1
selenium==3.141.0
six==1.15.0
urllib3==1.26.4
